backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
requests
python-dotenv
